Moved code into src
