--- runtime/__init__.py
@@ -1,1 +1,0 @@
-from .numeric_literals import underscore_literal_format
