+++ src/pypytranspy/runtime/__init__.py
@@ -1,0 +1,1 @@
+from pypytranspy.runtime.numeric_literals import underscore_literal_format
