+++ src/pypytranspy/runtime/numeric_literals.py
