+++ tests/transformations/remove_annotations_transformer_test.py
@@ -3,7 +3,7 @@ import ast
 import pytest
 import astor
 
-from transformations import RemoveAnnotationsTransformer
+from pypytranspy.transformations import RemoveAnnotationsTransformer
 
 
 @pytest.mark.parametrize(
