+++ tests/transformations/underscore_numeric_literals_transformer_test.py
@@ -3,7 +3,7 @@ import ast
 import pytest
 import astor
 
-from transformations import UnderscoreNumericLiteralsTransformer
+from pypytranspy.transformations import UnderscoreNumericLiteralsTransformer
 
 
 @pytest.mark.parametrize(
